+++ CHANGELOG.md
@@ -6,11 +6,14 @@ Simpleen is a [translation platform for i18n projects](https://simpleen.io/).
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.0.2] - 2021-01-19
+## [0.0.2] - 2021-01-22
 
 ### Second release
 
 - simpleen usage shows the current quota
+- simpleen upload saves adapted translations
+- [deprecated] simpleen lock - use simpleen upload instead
+- [fix] Catch error globally in BaseCommand
 
 ## [0.0.1] - 2020-10-06